budget: simplify renderSankeySvg extraction per /simplify review
- Move CategoryNode to a leaf module (category-node.ts) so the clean
  sankey-render module no longer imports its core type from the
  home-chart god module; drop the now-dead re-export.
- Restore the home-hydrate.ts producer fact into the
  TRANSACTIONS_APPENDED_EVENT JSDoc, lost when the inline comment was
  removed during extraction.
- Delete the dead firebase/firestore mock from sankey-render.test.ts;
  the module's import graph (d3-hierarchy, format.ts, type-only
  CategoryNode) never touches firestore.
